refactor(api): move associationName tests
