--- packages/api/tests/modules/association-name/__fixtures__/entity.ts
@@ -1,54 +1,0 @@
-import AssociationNameEntity from "../../../../src/modules/association-name/entities/AssociationNameEntity";
-
-export default [
-    {
-        lastUpdate: new Date("2022-07-21T00:26:31.635+02:00"),
-        name: "ASSOCIATION AURORE",
-        provider: "API SIRENE données ouvertes + API Répertoire des Associations (RNA)",
-        rna: "W931006748",
-        siren: "775684970",
-    },
-    {
-        lastUpdate: new Date("2022-09-02T03:23:45.181+02:00"),
-        name: "AVENIR ET JOIE - JOC",
-        provider: "API SIRENE données ouvertes + API Répertoire des Associations (RNA)",
-        rna: "W922001509",
-        siren: "775724842",
-    },
-    {
-        lastUpdate: { $date: "2022-01-19T00:00:00.000Z" },
-        name: "ORIN ÀBAJADE",
-        provider: "BASE RNA <Via API ASSO>",
-        rna: "W921009414",
-        siren: "884423187",
-    },
-    {
-        lastUpdate: { $date: "2015-12-21T00:00:00.000Z" },
-        name: "ASSOCIATION CONTRE CHAMPS",
-        provider: "BASE RNA <Via API ASSO>",
-        rna: "W913006937",
-        siren: "352535967",
-    },
-    {
-        lastUpdate: { $date: "2015-12-21T00:00:00.000Z" },
-        name: "ASSOCIATION CONTRE CHAMPS",
-        provider: "BASE RNA <Via API ASSO>",
-        rna: "W913006937",
-        siren: "352535967",
-    },
-
-    {
-        lastUpdate: new Date("2022-07-03T05:00:45.940+02:00"),
-        name: "GROUPEMENT D EMPLOYEURS PROFESSION SPORT LOISIRS",
-        provider: "API SIRENE données ouvertes + API Répertoire des Associations (RNA)",
-        rna: "W901000297",
-        siren: "502722978",
-    },
-    {
-        lastUpdate: new Date("2022-07-03T03:48:30.411+02:00"),
-        name: "LA CASERNE BASCULE",
-        provider: "API SIRENE données ouvertes + API Répertoire des Associations (RNA)",
-        rna: "W893006723",
-        siren: "899985618",
-    },
-] as AssociationNameEntity[];
