--- packages/api/tests/modules/association-name/associationName.service.test.ts
@@ -1,23 +1,0 @@
-import associationNameService from "../../../src/modules/association-name/associationName.service";
-import AssociationNameEntity from "../../../src/modules/association-name/entities/AssociationNameEntity";
-import associationNameRepository from "../../../src/modules/association-name/repositories/associationName.repository";
-
-describe("AssociationNameService", () => {
-    describe("getAllStartingWith", () => {
-        it("return an array AssociationNameEntity", async () => {
-            const INPUT = "";
-            const LAST_UPDATE = new Date();
-            const associationNameEntity = new AssociationNameEntity(
-                "W75000000",
-                "0000000000",
-                "FAKE NAME",
-                LAST_UPDATE,
-                null,
-            );
-            jest.spyOn(associationNameRepository, "findAllStartingWith").mockResolvedValue([associationNameEntity]);
-            const expected = [associationNameEntity];
-            const actual = await associationNameService.getAllStartingWith(INPUT);
-            expect(actual).toEqual(expected);
-        });
-    });
-});
